requirements.txt
fastapi
openai
python-dotenv
uvicorn
pydantic
python-multipart
sqlalchemy
psycopg2-binary
qdrant-client
supabase
streamlit
requests
fastembed
dash
dash-bootstrap-components